feat: update models factories and tables to account for money object

+++ database/migrations/2022_10_14_192143_create_accounts_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
         Schema::create((new Account)->getTable(), function (Blueprint $table) {
             $table->uuid('id');
             $table->string('name');
-            $table->float('balance')->default(0);
+            $table->bigInteger('balance')->default(0);
             $table->timestamps();
         });
     }
