+++ database/migrations/2022_10_14_201548_create_transactions_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->uuid('id');
             $table->foreignUuid('account_id');
             $table->string('name');
-            $table->float('amount');
+            $table->bigInteger('amount');
             $table->timestamps();
         });
     }
